Extract view pager animation logic into its own class.
